api/requirements.txt
fastapi
uvicorn

mysqlclient
databases
aiomysql
sqlalchemy

python-dotenv
